+++ templates/base.html
@@ -19,24 +19,24 @@
 <body>
     <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
         <div class="container">
-            <a class="navbar-brand" href="{{ url_for('index') }}">Scribe</a>
+            <a class="navbar-brand" href="{{ url_for('search.index') }}">Scribe</a>
             <div class="navbar-nav">
                 {% if current_user.is_authenticated %}
-                    <a class="nav-link" href="{{ url_for('write') }}">Write Story</a>
+                    <a class="nav-link" href="{{ url_for('stories.write') }}">Write Story</a>
                     <li class="nav-item">
-                        <a class="nav-link" href="{{ url_for('library') }}">
+                        <a class="nav-link" href="{{ url_for('library.view_library') }}">
                             <i class="fas fa-bookmark"></i> Library
                         </a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="{{ url_for('profile') }}">
+                        <a class="nav-link" href="{{ url_for('profile.view_profile') }}">
                             <i class="fas fa-user"></i> Profile
                         </a>
                     </li>
-                    <a class="nav-link" href="{{ url_for('logout') }}">Logout</a>
+                    <a class="nav-link" href="{{ url_for('auth.logout') }}">Logout</a>
                 {% else %}
-                    <a class="nav-link" href="{{ url_for('login') }}">Login</a>
-                    <a class="nav-link" href="{{ url_for('register') }}">Register</a>
+                    <a class="nav-link" href="{{ url_for('auth.login') }}">Login</a>
+                    <a class="nav-link" href="{{ url_for('auth.register') }}">Register</a>
                 {% endif %}
             </div>
         </div>
