+++ templates/login.html
@@ -15,5 +15,5 @@
         </div>
         <button type="submit" class="btn btn-primary">Login</button>
     </form>
-    <p class="mt-3">Don't have an account? <a href="{{ url_for('register') }}">Register</a></p>
+    <p class="mt-3">Don't have an account? <a href="{{ url_for('auth.register') }}">Register</a></p>
 {% endblock %}
