+++ templates/register.html
@@ -15,5 +15,5 @@
         </div>
         <button type="submit" class="btn btn-primary">Register</button>
     </form>
-    <p class="mt-3">Already have an account? <a href="{{ url_for('login') }}">Login</a></p>
+    <p class="mt-3">Already have an account? <a href="{{ url_for('auth.login') }}">Login</a></p>
 {% endblock %}
